Initial draft for Use-Cases as discussed with Lars
The use cases shall have a reference to the roles and activites described in the OpenChain Curriculum.
The use cases shall describe the "What". The User stories can be used to describe concrete features to be implemented in one sprint.
There is a description how use cases and user stories may be used together: https://www.visual-paradigm.com/guide/agile-software-development/user-story-vs-use-case/#integrate-use-case-user-story-and-story-map