Create Meeting-20231101.md
Minutes from Nov. 1.
Can committer please help that this is put in the right folder structure?  actually this should be under a folder "Meeting-20231101". I committed a file with that name yeterday, but didn't know how to make that be a folder. If you can correct and then put these minutes there, that would be great. Thanks.